Refactoring to using sub-driver
